Update to python 3.6
Versions prior to 3.7 are EoL, and the lack of access to
`from __future__ import annotations` in <=3.6 make it not worth working
around.

Pypi stats shows there is still some 3.6 utilization, but as breaking 2.7 support is typically
handled by a major version bump, and going from 3.6 -> 3.7 is trivial, the regular
maintaince of "a major bump happened" should be enough for them to push
to 3.7.