Clean up test runner
Test code doesn't need to manually construct a TestSuite and a
TextTestRunner, the unittest module has a discovery function that does
all this for you.

Delete all of the manual logic from tests.py, replace it with the two
lines to bring in the doctest unit tests, and update the makefile to
run the unittest discovery.